Moved the Jenkinsfile adder back to the task since the design changed and again it is only used in a single place. Renamed the JenkinsFiles helper class to more accurately describe it's usage.
